refactor(targets): decouple external plugin catalog assembly
