+++ crates/targets/src/manifest.rs
@@ -12,10 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::control_plane::external_target_plugin_installation;
 use crate::domain::TargetDomain;
-use crate::runtime::sidecar::SidecarPluginRuntime;
-use crate::runtime::sidecar_protocol::{SIDECAR_RUNTIME_PROTOCOL_VERSION, SidecarHandshake, SidecarPluginCapability};
 use rustfs_config::{
     AMQP_PASSWORD, AMQP_TLS_CLIENT_CERT, AMQP_TLS_CLIENT_KEY, KAFKA_TLS_CLIENT_CERT, KAFKA_TLS_CLIENT_KEY, MQTT_PASSWORD,
     MQTT_TLS_CLIENT_CERT, MQTT_TLS_CLIENT_KEY, MYSQL_DSN_STRING, MYSQL_TLS_CLIENT_CERT, MYSQL_TLS_CLIENT_KEY,
@@ -99,14 +96,6 @@ pub struct TargetPluginMarketplaceManifest {
     pub api_compatibility_version: &'static str,
     pub runtime_contract: TargetPluginExternalRuntimeContract,
     pub distribution: Option<TargetPluginDistributionManifest>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ExampleInstallableTargetPlugin {
-    pub manifest: TargetPluginMarketplaceManifest,
-    pub installation: crate::TargetPluginInstallation,
-    pub runtime: SidecarPluginRuntime,
-    pub valid_fields: Vec<String>,
 }
 
 const BUILTIN_PLUGIN_API_COMPATIBILITY_VERSION: &str = "rustfs.target-plugin.v1";
@@ -207,63 +196,6 @@ pub fn installable_target_marketplace_manifest(
     }
 }
 
-pub fn example_external_webhook_plugin() -> ExampleInstallableTargetPlugin {
-    let base = TargetPluginManifest {
-        plugin_id: "external:webhook-sidecar",
-        display_name: "Webhook Sidecar",
-        provider: "rustfs-labs",
-        version: "1.0.0",
-        target_type: "webhook",
-        supported_domains: &[TargetDomain::Notify],
-        secret_fields: &["auth_token"],
-    };
-    let manifest = installable_target_marketplace_manifest(
-        base,
-        TargetPluginEntrypointKind::Sidecar,
-        TargetPluginExternalRuntimeContract {
-            protocol_version: SIDECAR_RUNTIME_PROTOCOL_VERSION,
-            transport: TargetPluginRuntimeTransport::Grpc,
-        },
-        TargetPluginDistributionManifest {
-            artifacts: &[TargetPluginArtifactManifest {
-                artifact_id: "sidecar-linux-amd64",
-                target_triple: "x86_64-unknown-linux-gnu",
-                download_uri: "https://plugins.example.test/webhook-sidecar.tar.zst",
-                digest_sha256: "0123456789abcdef0123456789abcdef",
-                size_bytes: 8192,
-            }],
-        },
-    );
-
-    let handshake = SidecarHandshake {
-        protocol_version: SIDECAR_RUNTIME_PROTOCOL_VERSION.to_string(),
-        plugin_id: base.plugin_id.to_string(),
-        plugin_version: base.version.to_string(),
-        supported_domains: vec![TargetDomain::Notify],
-        capabilities: vec![
-            SidecarPluginCapability::HealthCheck,
-            SidecarPluginCapability::SendEvent,
-            SidecarPluginCapability::Shutdown,
-        ],
-    };
-    let mut runtime = SidecarPluginRuntime::new("grpc://127.0.0.1:50051", handshake);
-    runtime
-        .enable(base.plugin_id, TargetDomain::Notify)
-        .expect("example sidecar plugin handshake should validate");
-
-    ExampleInstallableTargetPlugin {
-        manifest,
-        installation: external_target_plugin_installation(
-            base.version,
-            "0123456789abcdef0123456789abcdef",
-            "sidecar-linux-amd64",
-            Some("2026-05-13T20:00:00Z".to_string()),
-        ),
-        runtime,
-        valid_fields: vec!["endpoint".to_string(), "auth_token".to_string()],
-    }
-}
-
 #[inline]
 fn builtin_plugin_id(target_type: &'static str) -> &'static str {
     match target_type {
@@ -286,7 +218,7 @@ mod tests {
         TargetPluginArtifactManifest, TargetPluginDistributionManifest, TargetPluginEntrypointKind,
         TargetPluginExternalRuntimeContract, TargetPluginMarketplaceManifest, TargetPluginPackaging,
         TargetPluginRuntimeTransport, builtin_target_manifest, builtin_target_marketplace_manifest,
-        example_external_webhook_plugin, installable_target_marketplace_manifest,
+        installable_target_marketplace_manifest,
     };
     use crate::domain::TargetDomain;
     use rustfs_config::{WEBHOOK_AUTH_TOKEN, WEBHOOK_CLIENT_CERT, WEBHOOK_CLIENT_KEY};
@@ -378,15 +310,5 @@ mod tests {
                 }],
             })
         );
-    }
-
-    #[test]
-    fn example_external_plugin_exposes_installation_and_runtime_metadata() {
-        let example = example_external_webhook_plugin();
-
-        assert_eq!(example.manifest.plugin_id, "external:webhook-sidecar");
-        assert_eq!(example.installation.install_state, crate::TargetPluginInstallState::Installed);
-        assert!(example.runtime.healthy);
-        assert_eq!(example.valid_fields, vec!["endpoint".to_string(), "auth_token".to_string()]);
     }
 }
